+++ src/vs/workbench/contrib/notebook/browser/services/notebookServiceImpl.ts
@@ -8,7 +8,7 @@
 import { usingPositronNotebooks } from '../../../positronNotebook/common/positronNotebookCommon.js';
 import { POSITRON_NOTEBOOK_ENABLED_KEY } from '../../../positronNotebook/common/positronNotebookConfig.js';
 import { IPYNB_VIEW_TYPE } from '../notebookBrowser.js';
-import { resolveNotebookSerializerOptions } from '../contrib/clean/cleanNotebooks.js';
+import { resolveNotebookSerializerOptions } from '../contrib/saveSettings/saveSettings.js';
 // --- End Positron ---
 import { localize } from '../../../../../nls.js';
 import { toAction } from '../../../../../base/common/actions.js';
